[160_common_interface], issue #160, renamed common_interface in tch_common_interface
