+++ lbh15/properties/lbe_thermochemical_properties/lbe_oxygen_limits.py
@@ -5,14 +5,14 @@ from typing import Union
 import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
-from ..common_interface import LowerLimitSaturationIronInterface
-from ..common_interface import LowerLimitSaturationChromiumInterface
-from ..common_interface import LowerLimitSaturationNickelInterface
-from ..common_interface import LowerLimitSaturationSiliconInterface
-from ..common_interface import LowerLimitSaturationAluminiumInterface
-from ..common_interface import LowerLimitChromiumInterface
-from ..common_interface import LowerLimitNickelInterface
-from ..common_interface import LowerLimitIronInterface
+from ..tch_common_interface import LowerLimitSaturationIronInterface
+from ..tch_common_interface import LowerLimitSaturationChromiumInterface
+from ..tch_common_interface import LowerLimitSaturationNickelInterface
+from ..tch_common_interface import LowerLimitSaturationSiliconInterface
+from ..tch_common_interface import LowerLimitSaturationAluminiumInterface
+from ..tch_common_interface import LowerLimitChromiumInterface
+from ..tch_common_interface import LowerLimitNickelInterface
+from ..tch_common_interface import LowerLimitIronInterface
 from .solubility_in_lbe import OxygenSolubility
 from .solubility_in_lbe import ChromiumSolubilityGosse2014
 from .solubility_in_lbe import ChromiumSolubilityCourouau2004
