+++ lbh15/properties/lbe_thermochemical_properties/lbe_thermochemical.py
@@ -6,10 +6,10 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from ..interface import PropertyInterface
-from ..common_interface import OxygenPartialPressureInterface
-from ..common_interface import MolarEnthalpyInterface
-from ..common_interface import MolarEntropyInterface
-from ..common_interface import GibbsFreeEnergyInterface
+from ..tch_common_interface import OxygenPartialPressureInterface
+from ..tch_common_interface import MolarEnthalpyInterface
+from ..tch_common_interface import MolarEntropyInterface
+from ..tch_common_interface import GibbsFreeEnergyInterface
 from ..lbe_properties import h
 from ..._commons import LBE_BOILING_TEMPERATURE as T_b0
 from ..._commons import LBE_MELTING_TEMPERATURE as T_m0
