+++ lbh15/properties/lead_thermochemical_properties/diffusivity_in_lead.py
@@ -4,9 +4,9 @@ from typing import List
 import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
-from lbh15.properties.interface import PropertyInterface
-from lbh15.properties.common_interface import OxygenDiffusivityInterface
-from lbh15.properties.common_interface import IronDiffusivityInterface
+from ..interface import PropertyInterface
+from ..tch_common_interface import OxygenDiffusivityInterface
+from ..tch_common_interface import IronDiffusivityInterface
 from ..._decorators import range_warning
